feat: allow any line format as (non-semantics) entry
