--- src/main/java/org/ingomohr/trac/in/TracReaderException.java
@@ -1,23 +1,0 @@
-package org.ingomohr.trac.in;
-
-public class TracReaderException extends Exception {
-
-    private static final long serialVersionUID = -7733494506820102231L;
-
-    public TracReaderException() {
-        super();
-    }
-
-    public TracReaderException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public TracReaderException(String message) {
-        super(message);
-    }
-
-    public TracReaderException(Throwable cause) {
-        super(cause);
-    }
-
-}
